+++ Sources/PowerSync/Kotlin/KotlinPowerSyncDatabaseImpl.swift
@@ -1,3 +1,4 @@
+import CSQLite
 import Foundation
 import PowerSyncKotlin
 
@@ -16,7 +17,12 @@ final class KotlinPowerSyncDatabaseImpl: PowerSyncDatabaseProtocol,
         dbFilename: String,
         logger: DatabaseLogger
     ) {
-        let factory = PowerSyncKotlin.DatabaseDriverFactory()
+        let rc = sqlite3_initialize()
+        if rc != 0 {
+            fatalError("Call to sqlite3_initialize() failed with \(rc)")
+        }
+
+        let factory = sqlite3DatabaseFactory(initialStatements: [])
         kotlinDatabase = PowerSyncDatabase(
             factory: factory,
             schema: KotlinAdapter.Schema.toKotlin(schema),
@@ -64,7 +70,8 @@ final class KotlinPowerSyncDatabaseImpl: PowerSyncDatabaseProtocol,
                 newClient: resolvedOptions.newClientImplementation,
                 userAgent: userAgent(),
                 loggingConfig: resolvedOptions.clientConfiguration?.requestLogger?.toKotlinConfig()
-            )
+            ),
+            appMetadata: resolvedOptions.appMetadata
         )
     }
 
@@ -89,9 +96,10 @@ final class KotlinPowerSyncDatabaseImpl: PowerSyncDatabaseProtocol,
         try await kotlinDatabase.disconnect()
     }
 
-    func disconnectAndClear(clearLocal: Bool = true) async throws {
+    func disconnectAndClear(clearLocal: Bool, soft: Bool) async throws {
         try await kotlinDatabase.disconnectAndClear(
-            clearLocal: clearLocal
+            clearLocal: clearLocal,
+            soft: soft
         )
     }